*) Code from Tiv - Added: Experimental code for support of the new emblems.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/branch/1.x@241 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/keeps/Gameobjects/Guards/GameKeepGuard.cs
@@ -675,7 +675,7 @@ result == GameLiving.eAttackResult.Parried)
 			if (this.Inventory == null)
 				return;
 
-			ushort emblem = 0;
+			int emblem = 0;
 			if (guild != null)
 				emblem = guild.theGuildDB.Emblem;
 			InventoryItem lefthand = this.Inventory.GetItem(eInventorySlot.LeftHandWeapon);
